Rename Google map component files
