+++ src/components/Map/DynamicGoogleMap.js
