+++ src/components/Map/StaticGoogleMap.js
